requirements.txt
ImageHash==4.1.0
numpy==1.19.1
Pillow==7.2.0
PyWavelets==1.1.1
scipy==1.5.2
six==1.15.0
